requirements.txt
streamlit==1.36.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.0
tenacity==8.5.0
tavily-python==0.4.0
rapidfuzz==3.9.6
pdfminer.six==20240706
